requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.9.1
    # via
    #   django
    #   django-allauth
certifi==2025.8.3
    # via requests
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.4.3
    # via requests
colorama==0.4.6
    # via pytest
coverage[toml]==7.10.3
    # via pytest-cov
distlib==0.4.0
    # via virtualenv
dj-database-url==3.0.1
    # via -r requirements.in
dj-rest-auth==7.0.1
    # via -r requirements.in
django==5.2.5
    # via
    #   -r requirements.in
    #   dj-database-url
    #   dj-rest-auth
    #   django-allauth
    #   djangorestframework
    #   djangorestframework-simplejwt
django-allauth==65.11.0
    # via -r requirements.in
djangorestframework==3.16.1
    # via
    #   -r requirements.in
    #   dj-rest-auth
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.5.1
    # via -r requirements.in
factory-boy==3.3.3
    # via -r requirements.in
faker==37.5.3
    # via factory-boy
filelock==3.18.0
    # via virtualenv
identify==2.6.13
    # via pre-commit
idna==3.10
    # via requests
iniconfig==2.1.0
    # via pytest
nodeenv==1.9.1
    # via pre-commit
packaging==25.0
    # via pytest
pillow==11.3.0
    # via -r requirements.in
platformdirs==4.3.8
    # via virtualenv
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
pre-commit==4.3.0
    # via -r requirements.in
psycopg2-binary==2.9.10
    # via -r requirements.in
pygments==2.19.2
    # via pytest
pyjwt==2.10.1
    # via djangorestframework-simplejwt
pytest==8.4.1
    # via
    #   -r requirements.in
    #   pytest-cov
    #   pytest-django
pytest-cov==6.2.1
    # via -r requirements.in
pytest-django==4.11.1
    # via -r requirements.in
python-decouple==3.8
    # via -r requirements.in
pyyaml==6.0.2
    # via pre-commit
requests==2.32.5
    # via -r requirements.in
ruff==0.12.8
    # via -r requirements.in
sqlparse==0.5.3
    # via django
tzdata==2025.2
    # via
    #   django
    #   faker
urllib3==2.5.0
    # via requests
virtualenv==20.33.1
    # via pre-commit
